pyproject.toml
[tool.poetry]
name = "jobson"
version = "0.1.0"
description = ""
authors = ["Pedro Balbino <phsb5321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pandas = "^2.1.4"
numpy = "^1.26.3"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
flask = "^3.0.0"
apscheduler = "^3.10.4"
flask-restx = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
